Fixed game level spinner width bug
The bug was happening because of how the Spinner's width measurement is implemented. It only measures 15 items at a time and sets the width of that section based on the widest item from within it. To overcome this, I had to add spaces to the digits in order to make their length as a string be equal to that of the word "Select".