Fixed bug in ghost piece feature.
Fixed a bug that was caused by the sortedGridKeys list to be empty if the ghost wasn't enabled. This meant that if the ghost was suddenly enabled mid-game and the grid wasn't empty, the ghost would be drawn in positions that made it collide with squares in the grid. The solution was to simply sort the keys each time the grid is redrawn, regardless of whether the ghost is enabled or not.